fix: Resolve iOS API type mismatches after main merge
- MediaPlayerController: Fix NSData initialization using NSMutableData.appendBytes
- SystemAppearance: Use UIUserInterfaceStyle enum instead of raw Long values
- DataChannelWrapper: Fix NSData initialization in toNSData() helper

These corrections ensure compatibility with the Kotlin/Native iOS interop layer
and resolve compilation errors introduced during the merge.

+++ composeApp/src/iosMain/kotlin/io/music_assistant/client/player/MediaPlayerController.ios.kt
@@ -90,7 +90,7 @@ actual class MediaPlayerController actual constructor(platformContext: PlatformC
     }
 
     actual fun getCurrentSystemVolume(): Int {
-        return (AVAudioSession.sharedInstance().outputVolume * 100).toInt()
+        return (AVAudioSession.sharedInstance().outputVolume.toFloat() * 100).toInt()
     }
     
     // Now Playing (Control Center / Lock Screen)
